config project

+++ src/main/java/com/project/SafetyNet/service/FirestationService.java
@@ -1,8 +1,11 @@
 package com.project.SafetyNet.service;
 
+import com.project.SafetyNet.exception.RessourceNotFoundException;
 import com.project.SafetyNet.repository.FirestationRepository;
 import com.project.SafetyNet.model.Firestation;
 import lombok.Data;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
@@ -12,25 +15,41 @@ import java.util.List;
 public class FirestationService {
     @Autowired
     private FirestationRepository firestationRepository;
+    @Autowired
+    static final Logger logger = LogManager.getLogger(FirestationService.class);
+
     private String address;
 
     public List<Firestation> getAllFirestation(){
         return firestationRepository.findAllFirestation();
     }
 
-    public Firestation addFirestation(Firestation firestation) {
-        firestationRepository.addFirestation(firestation);
+    public Firestation addFirestation(Firestation firestation) throws RessourceNotFoundException {
+        if(firestation!=null) {
+            firestationRepository.addFirestation(firestation);
+        }else {
+            logger.error("firestation is empty");
+            throw new RessourceNotFoundException("firestation don't exist");
+
+        }
         return firestation;
     }
 
-    public Firestation updateFirestation( String address,Firestation firestationToUpdate) {
-
-        return firestationRepository.updateFirestation(address,firestationToUpdate);
-
+    public Firestation updateFirestation( String address,Firestation firestationToUpdate) throws RessourceNotFoundException {
+        if(address!=null) {
+            return firestationRepository.updateFirestation(address, firestationToUpdate);
+        }else {
+            logger.error("the" + firestationToUpdate + "don't exist");
+            throw new RessourceNotFoundException("firestation to update don't exist");
+        }
     }
 
     public void deleteFirestation(String address) {
-        firestationRepository.deleteFirestation(address);
+        if(address!=null) {
+            firestationRepository.deleteFirestation(address);
+        }else{
+            logger.error("firestation don't exist");
+    }
     }
